migrate activity/handles-test to testing-library and ts
